Change token and auth middleware to be more flexible
Can create options when creating instance of midd-
leware functions so user of function can choose
whether to throw errors to the express error hand-
ler or not.

This means an unauthenticated user will be able to
access all routes even if the token fails to verify,
or getUser fails, unless that route is protected
by auth isAuth or isAdmin middlewares, or the
showError option is set to true.