Move admin routes from regular routers to behind /admin router
So all the /admin routes can only be made from the
admin client domain, and can easily check that an
admin is logged in for all admin routes.

Also so that only an admin can log in from the ad-
min client domain. If a non-admin tries to log in,
it will reject the request.

The routes that all required admin permissions have
been removed from the standard routes and moved to
the admin equivalent.

This has the benefit of a lot less checking of whether
the user is an admin on individual routes. Now it is
just in one place.

Any routes that don't need admin permissions have
been added to the admin route anyway. So for example,
/comments and /admin/comments do basically the same
stuff. But I included them on the /admin router a-
nyway because it would be easy to make a mistake
when creating the admin client if some of the routes
were /admin/etc and others were just e.g. /comments.