Store refresh token in httpOnly secure cookie instead of http response auth header
As far as I can tell, there is no way on the clientside
to extract the refresh token from a http response auth
header except to use javascript. All sources indicate
that giving js access to a refresh token is very bad.
If some malicious code (from any of the millions of
packages we are all using) got access to a refresh token,
yeah that would be bad.

So I am setting the refresh token in a secure cookie.
All the access tokens will be sent to the server in the
http request auth header (js required for that), but when
a new access token is requested, the refresh token will
be sent over automatically in the cookie.

All instances of tokens being required in the url query
params will be moved to using a header auth token, except
the account confirmation and password reset routes.