Include id in jwt so client can decode and use to fetch only the logged in user's comments
Otherwise, I would have to create a new server route
for fetching only the logged in user's comments,
which would check the provided access token and verify
it has not been tampered with, and then return the
user's comments.

Since all comments are public, it is not a problem
that the client could easily change the id and fetch
different comments.